Refactor: Consistent naming/view structure for summary views
